+++ coderedcms/templates/coderedcms/blocks/pagelist_block.html
@@ -1,12 +1,12 @@
 {% extends 'coderedcms/blocks/base_block.html' %}
-{% load wagtailcore_tags wagtailimages_tags %}
+{% load wagtailcore_tags %}
 
 {% block block_render %}
 
 <ul>
     {% for page in pages %}
     {% with page=page.specific %}
-    <li><a href="{{page.url}}">
+    <li><a href="{% pageurl page.url %}">
         {{page.title}} {% if self.show_preview %}<small class="text-muted">– {{page.body_preview}}</small>{% endif %}
     </a></li>
     {% endwith %}
